+++ user/tests/factories.py
@@ -8,8 +8,8 @@ class UserFactory(factory.django.DjangoModelFactory):
         model = User
         django_get_or_create = ("username",)
 
-    username = factory.Faker("user_name")
-    email = factory.Faker("email")
+    username = factory.Sequence(lambda n: f"user{n}")
+    email = factory.LazyAttribute(lambda obj: f"{obj.username}@example.com")
 
 
 class UserWithPasswordFactory(UserFactory):